cart

+++ view/Cart.php
@@ -9,11 +9,11 @@
     <h1><center>Panier<center></h1>
     </br>
 	<?php
-
+if(isset($_SESSION["id"])){
 	$reponse=$bdd->query('select * from products p
 INNER JOIN order_products op ON p.id = op.product_id
 WHERE op.order_id = \''.$_SESSION["id"].'\'');
-
+}
 $somme = 0;
 $prix = 0;
 	?>
@@ -29,24 +29,27 @@ $prix = 0;
         </tr></thead>
 	<tbody>
 	
-		<?php while($donnees=$reponse->fetch()){
-			?>
-			<tr>
-            <td> <?php echo $donnees['name'] ?></td>
-            <td><image src=<?php echo $donnees['lien_image'] ?> width="200px"></td>
-            <td><?php echo $donnees['cutiness'] ?></td>
-			<td><?php echo $donnees['quantity'] ?></td>
-			<td><?php echo $donnees['unit_price']?></td>
-			<td><?php echo( $donnees['quantity']* $donnees['unit_price']) ?></td>
-            <td><form method="post">
-			<input type="hidden" name="delete" value=<?php echo $donnees["product_id"] ?> >
-			<input type="submit" value="delete" >
-			</form> </td> 
-            </tr>
+		<?php
+		if(isset($_SESSION["id"])){
+			while($donnees=$reponse->fetch()){
+				?>
+				<tr>
+				<td> <?php echo $donnees['name'] ?></td>
+				<td><image src=<?php echo $donnees['lien_image'] ?> width="200px"></td>
+				<td><?php echo $donnees['cutiness'] ?></td>
+				<td><?php echo $donnees['quantity'] ?></td>
+				<td><?php echo $donnees['unit_price']?></td>
+				<td><?php echo( $donnees['quantity']* $donnees['unit_price']) ?></td>
+				<td><form method="post">
+				<input type="hidden" name="delete" value=<?php echo $donnees["product_id"] ?> >
+				<input type="submit" value="delete" >
+				</form> </td> 
+				</tr>
 
-			<?php
-			$somme+=$donnees['quantity'];
-			$prix+=$donnees['quantity']*$donnees['unit_price'];
+				<?php
+				$somme+=$donnees['quantity'];
+				$prix+=$donnees['quantity']*$donnees['unit_price'];
+			}
 		}?>
 		<td> total</td>
 		<td></td>
@@ -55,7 +58,9 @@ $prix = 0;
 		<td> prix total du panier</td>
 		<td><?php echo $prix?></td>
 		<td><form method="post">
+		<?php if(isset($_SESSION["id"])){ ?>
 			<input type="hidden" name="delete_all" value=<?php echo $donnees["order_id"] ?> >
+			<?php } ?>
 			<input type="submit" value="vider le panier" >
 			</form></td>
 	</tbody>
